Component to override all the registered traslation keys
This component tracks every registered ResourceBundleLoader,
organizes them by bundle.symbolic.name and resource.bundle.base.name and
takes the one with the highest ranking (the expected active one). Then
registers an aggregate of the properties coming in the very bundle on
top of the active ResourceBundleLoader using its very properties thus
effectively overriding it.

This component does not mix one component keys with each others keys.

The caveat is that if some @Component is bound to a ResourceBundleLoader
using RELUCTANT policyOption (the default) it might not get the new keys
until the service is refreshed (or the portal restarted)